+++ apps/web/app/(app)/bulk-unsubscribe/BulkUnsubscribeSection.tsx
@@ -22,9 +22,7 @@ import { usePremium } from "@/components/PremiumAlert";
 import {
   useNewsletterFilter,
   useBulkUnsubscribeShortcuts,
-  useBulkUnsubscribeButton,
-  useBulkApproveButton,
-} from "@/app/(app)/bulk-unsubscribe/common";
+} from "@/app/(app)/bulk-unsubscribe/hooks";
 import BulkUnsubscribeSummary from "@/app/(app)/bulk-unsubscribe/BulkUnsubscribeSummary";
 import { useStatLoader } from "@/providers/StatLoaderProvider";
 import { usePremiumModal } from "@/app/(app)/premium/PremiumModal";
@@ -41,8 +39,7 @@ import { Card } from "@/components/ui/card";
 import { ShortcutTooltip } from "@/app/(app)/bulk-unsubscribe/ShortcutTooltip";
 import { SearchBar } from "@/app/(app)/bulk-unsubscribe/SearchBar";
 import { useToggleSelect } from "@/hooks/useToggleSelect";
-import { Button } from "@/components/ui/button";
-import { ButtonLoader } from "@/components/Loading";
+import { BulkActions } from "@/app/(app)/bulk-unsubscribe/BulkActions";
 
 type Newsletter = NewsletterStatsResponse["newsletters"][number];
 
@@ -155,96 +152,13 @@ export function BulkUnsubscribeSection({
     );
   });
 
-  const { bulkUnsubscribeLoading, onBulkUnsubscribe } =
-    useBulkUnsubscribeButton({
-      hasUnsubscribeAccess,
-      mutate,
-      posthog,
-      refetchPremium,
-    });
-
-  const { bulkApproveLoading, onBulkApprove } = useBulkApproveButton({
-    mutate,
-    posthog,
-  });
-
   return (
     <>
       {!isMobile && <BulkUnsubscribeSummary />}
       <Card className="mt-0 md:mt-4">
         <div className="items-center justify-between px-2 pt-2 sm:px-4 md:flex">
           {Array.from(selected.values()).filter(Boolean).length > 0 ? (
-            <div className="flex items-center space-x-1.5">
-              <div>
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={() =>
-                    onBulkUnsubscribe(
-                      Array.from(selected.entries()).map(([name, value]) => ({
-                        name,
-                        value,
-                      })),
-                    )
-                  }
-                  disabled={bulkUnsubscribeLoading}
-                >
-                  {bulkUnsubscribeLoading && <ButtonLoader />}
-                  Unsubscribe
-                </Button>
-              </div>
-              {/* <div>
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={rejectSelected}
-                  disabled={isApproving || isRejecting}
-                >
-                  {isRejecting && <ButtonLoader />}
-                  Auto Archive
-                </Button>
-              </div> */}
-              <div>
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={() =>
-                    onBulkApprove(
-                      Array.from(selected.entries()).map(([name, value]) => ({
-                        name,
-                        value,
-                      })),
-                    )
-                  }
-                  disabled={bulkApproveLoading}
-                >
-                  {bulkApproveLoading && <ButtonLoader />}
-                  Approve
-                </Button>
-              </div>
-              {/* <div>
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={rejectSelected}
-                  disabled={isApproving || isRejecting}
-                >
-                  {isRejecting && <ButtonLoader />}
-                  Archive
-                </Button>
-              </div> */}
-              {/* <div>
-                <Button
-                  size="sm"
-                  variant="outline"
-                  onClick={rejectSelected}
-                  disabled={isApproving || isRejecting}
-                >
-                  {isRejecting && <ButtonLoader />}
-                  Delete
-                </Button>
-              </div> */}
-            </div>
+            <BulkActions selected={selected} mutate={mutate} />
           ) : (
             <Title className="hidden md:block">
               Bulk unsubscribe from emails
